:recycle: refactor & tidy up existing extraction module
