+++ lib/mindee/extraction/common.rb
