+++ lib/mindee/extraction/common/extracted_image.rb
