+++ lib/mindee/extraction/common/image_extractor.rb
