+++ lib/mindee/extraction/multi_receipts_extractor.rb
