+++ lib/mindee/extraction/tax_extractor/ocr_extractor.rb
